refactor: criando serviço para a collection de Links e composable useFolders
