+++ src/pages/Links/index.vue
@@ -1,19 +1,11 @@
 <script setup lang="ts">
 import { ref, onMounted, watch } from 'vue'
 import { useRoute } from 'vue-router'
-import {
-  collection,
-  doc,
-  getDoc,
-  getDocs,
-  orderBy,
-  query,
-  where,
-} from 'firebase/firestore'
 
 import { Link } from '@/interfaces/link'
 import { Folder } from '@/interfaces/folder'
-import { db } from '@/lib/firebase'
+import { folderService } from '@/services/folder'
+import { linkService } from '@/services/link'
 
 import Loader from '@/components/Loader.vue'
 import FolderInfo from './components/FolderInfo.vue'
@@ -46,18 +38,8 @@ function updateFolder(name: string) {
 
 async function fetchFolderData(folderId: string) {
   try {
-    const folderRef = doc(db, 'folders', folderId)
-    const data = await getDoc(folderRef)
-
-    if (!data.exists()) {
-      return
-    }
-
-    folder.value = {
-      id: data.id,
-      name: data.data().name,
-      createdAt: data.data().createdAt.toDate(),
-    }
+    const data = await folderService.findById(folderId)
+    folder.value = data
   } catch {
     throw new Error('Não foi possível buscar as informações pasta')
   }
@@ -65,25 +47,8 @@ async function fetchFolderData(folderId: string) {
 
 async function fetchLinks(folderId: string) {
   try {
-    const q = query(
-      collection(db, 'links'),
-      orderBy('createdAt', 'asc'),
-      where('folderId', '==', folderId)
-    )
-
-    const querySnapshot = await getDocs(q)
-    const linksData: Link[] = []
-
-    querySnapshot.forEach((doc) => {
-      linksData.push({
-        id: doc.id,
-        title: doc.data().title,
-        url: doc.data().url,
-        folderId: doc.data().folderId,
-      })
-    })
-
-    links.value = linksData
+    const data = await linkService.findAll(folderId)
+    links.value = data
   } catch {
     throw new Error('Não foi possível buscar os links')
   }
