+++ src/services/folder.ts
@@ -3,10 +3,12 @@ import {
   addDoc,
   collection,
   doc,
+  getDoc,
   updateDoc,
 } from 'firebase/firestore'
 
 import { db } from '@/lib/firebase'
+import { Folder } from '@/interfaces/folder'
 
 interface FolderData {
   name: string
@@ -25,5 +27,20 @@ export const folderService = {
   update: async (id: string, name: string) => {
     const folderRef = doc(db, 'folders', id)
     return updateDoc(folderRef, { name })
+  },
+
+  findById: async (id: string): Promise<Folder | null> => {
+    const folderRef = doc(db, 'folders', id)
+    const data = await getDoc(folderRef)
+
+    if (!data.exists()) {
+      return null
+    }
+
+    return {
+      id: data.id,
+      name: data.data().name,
+      createdAt: data.data().createdAt.toDate(),
+    }
   },
 }
